backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
alembic>=1.12.1
cryptography>=41.0.7
Faker>=22.5.1
python-multipart>=0.0.6
email-validator>=2.1.0.post1
typing-extensions>=4.8.0
